Actualizacion excepciones administrador

+++ src/main/java/com/example/hulkstore/Controller/AdminController.java
@@ -31,7 +31,7 @@ public class AdminController {
         }
     }
 
-    @GetMapping("/verAdmins/{adminId}")//Listar usuarios
+    @GetMapping("/verAdmin/{adminId}")//Listar usuarios
     public ResponseEntity<AdminDTO> getAdminById(@PathVariable("adminId") Long adminId) {
         try{
             Optional<AdminDTO> adminOptional = adminService.getAdminById(adminId);
@@ -47,12 +47,12 @@ public class AdminController {
     }
 
     @PostMapping("/addAdmin")//Agregar admin
-    public void addAdmin(@RequestBody Admin admin) {
+    public ResponseEntity<String> addAdmin(@RequestBody Admin admin) {
         try {
             adminService.addAdmin(admin);
-            ResponseEntity.ok("Admin añadido correctamente");
+            return ResponseEntity.ok("Admin añadido correctamente");
         } catch (Exception e) {
-            ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .body("Error al agregar Administrador");
         }
     }
@@ -60,9 +60,16 @@ public class AdminController {
     @PutMapping("/updateAdmin/{adminId}")//Actualizar admin
     public ResponseEntity<String> updateAdmin(@PathVariable("adminId") Long adminId, @RequestBody Admin admin) {
         try {
-            admin.setAdminId(adminId);
-            adminService.updateAdmin(admin);
-            return  ResponseEntity.ok("Administrado actualizado exitosamente");
+            Optional<AdminDTO> optionalAdmin = adminService.getAdminById(adminId);
+            if (optionalAdmin.isPresent()) {
+                admin.setAdminId(adminId);
+                adminService.updateAdmin(admin);
+                return  ResponseEntity.ok("Administrado actualizado exitosamente");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body("No existe administrador con la id: " + adminId);
+            }
+
         } catch (Exception e) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                     .body("Error al actualizar el Administrador");
@@ -71,8 +78,14 @@ public class AdminController {
     @DeleteMapping("/deleteAdmin/{adminId}")//Borrar admin
     public ResponseEntity<String> deleteAdminById(@PathVariable("adminId") Long adminId) {
         try {
-            adminService.deleteAdminById(adminId);
-            return ResponseEntity.ok("Administrador eliminado correctamente");
+            Optional<AdminDTO> optionalAdmin = adminService.getAdminById(adminId);
+            if (optionalAdmin.isPresent()){
+                adminService.deleteAdminById(adminId);
+                return ResponseEntity.ok("Administrador eliminado correctamente");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                        .body("No existe administrador con la id: " + adminId);
+            }
         } catch (Exception e) {
          return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
                  .body("Error al eliminar administrador");
